+++ migrate-cms-to-cds/src/main/java/org/acumos/cds/migrate/MigrateCmsToCdsApp.java
@@ -20,6 +20,7 @@
 package org.acumos.cds.migrate;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.lang.invoke.MethodHandles;
 import java.util.List;
@@ -28,10 +29,13 @@ import org.acumos.cds.client.CommonDataServiceRestClientImpl;
 import org.acumos.cds.client.ICommonDataServiceRestClient;
 import org.acumos.cds.domain.MLPDocument;
 import org.acumos.cds.domain.MLPRevisionDescription;
+import org.acumos.cds.domain.MLPSiteConfig;
+import org.acumos.cds.domain.MLPSiteContent;
 import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPSolutionRevision;
 import org.acumos.cds.migrate.client.CMSReaderClient;
 import org.acumos.cds.migrate.client.CMSWorkspace;
+import org.acumos.cds.migrate.domain.CMSDescription;
 import org.acumos.cds.migrate.domain.CMSNameList;
 import org.acumos.cds.migrate.domain.CMSRevisionDescription;
 import org.acumos.cds.transport.RestPageRequest;
@@ -42,6 +46,14 @@ import org.acumos.nexus.client.data.UploadArtifactInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.client.HttpStatusCodeException;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.BooleanNode;
+import com.fasterxml.jackson.databind.node.IntNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.databind.node.TextNode;
 
 /**
  * Migrates the following data from CMS to CDS and Nexus:
@@ -62,11 +74,12 @@ public class MigrateCmsToCdsApp {
 
 	// I think this will grow
 	private static final String specialCharRegex = "[!@#$%^&*()<>{}|:?+=.\\s]+";
+
 	/**
 	 * Migrates data.
 	 * 
 	 * @param args
-	 *            Ignored
+	 *                 Ignored
 	 */
 	public static void main(String[] args) {
 
@@ -75,6 +88,7 @@ public class MigrateCmsToCdsApp {
 		int migrPicSucc = 0, migrPicFail = 0;
 		int migrDescSucc = 0, migrDescFail = 0;
 		int migrDocSucc = 0, migrDocFail = 0;
+		int globalContentSucc = 0, globalContentFail = 0;
 
 		MigrateProperties props = null;
 		try {
@@ -109,6 +123,8 @@ public class MigrateCmsToCdsApp {
 			return;
 		}
 
+		// PHASE 1: DOCUMENTS, PICTURES ETC.
+
 		final int pageSize = 100;
 		for (int page = 0;; ++page) {
 			RestPageRequest request = new RestPageRequest(page, pageSize);
@@ -124,13 +140,13 @@ public class MigrateCmsToCdsApp {
 					logger.info("Solution {} has no CMS image", s.getSolutionId());
 				} else {
 					String imgName = solNames.getResponse_body().get(0);
-					if (s.getPicture() != null) {
+					if (cdsClient.getSolutionPicture(s.getSolutionId()) != null) {
 						// CDS already has it
 						logger.info("Solution {} image {} already migrated", s.getSolutionId(), imgName);
 					} else {
 						logger.info("Migrating solution {} image: {}", s.getSolutionId(), imgName);
 						byte[] cmsImage = cmsClient.getSolutionImage(s.getSolutionId(), imgName);
-						s.setPicture(cmsImage);
+						cdsClient.saveSolutionPicture(s.getSolutionId(), cmsImage);
 						try {
 							cdsClient.updateSolution(s);
 							++migrPicSucc;
@@ -202,7 +218,7 @@ public class MigrateCmsToCdsApp {
 									++migrDocFail;
 								} else {
 									// Produce a clean basename for Nexus by replacing special characters
-									final String nexusDocBase = cmsDocParts[0].replaceAll(specialCharRegex, "-");	
+									final String nexusDocBase = cmsDocParts[0].replaceAll(specialCharRegex, "-");
 									final String nexusDocSuffix = cmsDocParts[1];
 									final String nexusDocName = nexusDocBase + "." + nexusDocSuffix;
 									if (findDocNameInCdsList(nexusDocName, cdsRevDocs)) {
@@ -235,7 +251,7 @@ public class MigrateCmsToCdsApp {
 											} catch (Exception ex) {
 												logger.error(
 														"Failed to upload revision {} doc base {} doc suffix as nexus artifact; exception follows",
-														r.getRevisionId(), nexusDocBase,  nexusDocSuffix);
+														r.getRevisionId(), nexusDocBase, nexusDocSuffix);
 												logger.error("Exception in upload", ex);
 												++migrDocFail;
 											}
@@ -280,21 +296,211 @@ public class MigrateCmsToCdsApp {
 
 		} // for page
 
+		// PHASE 2: GLOBAL SITE CONTENT
+
+		final String keyCobrandLogo = "global.coBrandLogo";
+		final String keyContactInfo = "global.footer.contactInfo";
+		final String keyTermsCondition = "global.termsCondition";
+
+		byte[] coBrandLogo = cmsClient.getCoBrandLogo();
+		if (coBrandLogo == null || coBrandLogo.length == 0) {
+			logger.error("Source CMS has no co-brand logo, continuing");
+			++globalContentFail;
+		} else {
+			MLPSiteContent cdsCoBrandLogo = cdsClient.getSiteContent(keyCobrandLogo);
+			if (cdsCoBrandLogo != null && cdsCoBrandLogo.getContentValue().length > 0) {
+				logger.info("Target CDS already has co-brand logo, continuing");
+			} else {
+				try {
+					logger.info("Creating co-brand logo in CDS");
+					cdsCoBrandLogo = new MLPSiteContent(keyCobrandLogo, coBrandLogo, "image/jpg");
+					cdsClient.createSiteContent(cdsCoBrandLogo);
+					++globalContentSucc;
+				} catch (HttpStatusCodeException ex) {
+					logger.error("Failed to create co-brand logo {}; server response {}", cdsCoBrandLogo,
+							ex.getResponseBodyAsString());
+					++globalContentFail;
+				}
+			}
+		}
+
+		CMSDescription cmsFootCi = cmsClient.getFooterContactInfo();
+		if (cmsFootCi == null || cmsFootCi.getDescription() == null || cmsFootCi.getDescription().isEmpty()) {
+			logger.info("Source CMS has no footer contact info, continuing");
+		} else {
+			MLPSiteContent cdsFootCi = cdsClient.getSiteContent(keyContactInfo);
+			if (cdsFootCi != null && cdsFootCi.getContentValue().length > 0) {
+				logger.info("Target CDS already has footer contact, continuing");
+			} else {
+				try {
+					logger.info("Creating footer contact site content in CDS");
+					cdsFootCi = new MLPSiteContent(keyContactInfo, cmsFootCi.getDescription().getBytes(), "text/html");
+					cdsClient.createSiteContent(cdsFootCi);
+					++globalContentSucc;
+				} catch (HttpStatusCodeException ex) {
+					logger.error("Failed to create footer contact {}; server response {}", cdsFootCi,
+							ex.getResponseBodyAsString());
+					++globalContentFail;
+				}
+			}
+		}
+		CMSDescription cmsFootTc = cmsClient.getFooterTermsConditions();
+		if (cmsFootTc == null || cmsFootTc.getDescription() == null || cmsFootTc.getDescription().isEmpty()) {
+			logger.info("Source CMS has no footer t&c, continuing");
+		} else {
+			MLPSiteContent cdsFootTc = cdsClient.getSiteContent(keyTermsCondition);
+			if (cdsFootTc != null && cdsFootTc.getContentValue().length > 0) {
+				logger.info("Target CDS already has footer T&C, continuing");
+			} else {
+				try {
+					logger.info("Creating footer T&C site content in CDS");
+					cdsFootTc = new MLPSiteContent(keyTermsCondition, cmsFootTc.getDescription().getBytes(),
+							"text/html");
+					cdsClient.createSiteContent(cdsFootTc);
+					++globalContentSucc;
+				} catch (HttpStatusCodeException ex) {
+					logger.error("Failed to create footer T&C {}; server response {}", cdsFootTc,
+							ex.getResponseBodyAsString());
+					++globalContentFail;
+				}
+			}
+		}
+
+		MLPSiteConfig topCarouselConfig = cdsClient.getSiteConfig("carousel_config");
+		if (topCarouselConfig == null) {
+			logger.warn("Failed to find top carousel config");
+		} else {
+			try {
+				String revisedConfig = migrateCarouselConfig(cmsClient, cdsClient, typeTopBg, typeTopIg, "top",
+						topCarouselConfig.getConfigValue());
+				topCarouselConfig.setConfigValue(revisedConfig);
+				// TODO cdsClient.updateSiteConfig(topCarouselConfig);
+				++globalContentSucc;
+			} catch (Exception ex) { //
+				logger.error("Failed to migrate top carousel, exception {}", ex.toString());
+				++globalContentFail;
+			}
+		}
+		MLPSiteConfig eventCarouselConfig = cdsClient.getSiteConfig("event_carousel");
+		if (eventCarouselConfig == null) {
+			logger.warn("Failed to find event carousel config");
+		} else {
+			try {
+				String revisedConfig = migrateCarouselConfig(cmsClient, cdsClient, typeEventBg, typeEventIg, "event",
+						eventCarouselConfig.getConfigValue());
+				eventCarouselConfig.setConfigValue(revisedConfig);
+				cdsClient.updateSiteConfig(eventCarouselConfig);
+				++globalContentSucc;
+			} catch (Exception ex) { //
+				logger.error("Failed to migrate event carousel, exception {}", ex.toString());
+				++globalContentFail;
+			}
+		}
+
 		logger.info("Migration statistics:");
 		logger.info("Solutions checked: {}", solCount);
 		logger.info("Revisions checked: {}", revCount);
 		logger.info("Pictures migrated: {} success, {} fail", migrPicSucc, migrPicFail);
 		logger.info("Descriptions migrated: {} success, {} fail", migrDescSucc, migrDescFail);
 		logger.info("Documents migrated: {} success, {} fail", migrDocSucc, migrDocFail);
+		logger.info("Global items migrated: {} success, {} fail", globalContentSucc, globalContentFail);
+
+	}
+
+	// Nightmare of different paths in CMS
+	private final static String typeTopBg = "carousel_background";
+	private final static String typeTopIg = "carousel_infoGraphic";
+	private final static String typeEventBg = "event_carousel_bg";
+	private final static String typeEventIg = "event_carousel_ig";
+
+	// Tags for CDS site config
+	// OLD
+	private final static String bgImageUrl = "bgImageUrl";
+	private final static String infoImageUrl = "infoImageUrl";
+	// NEW
+	private final static String uniqueKey = "uniqueKey";
+	private final static String bgImgKey = "bgImgKey";
+	private final static String infoImgKey = "infoImgKey";
+	private final static String hasInfoGraphic = "hasInfoGraphic";
+
+	/**
+	 * Migrates carousel images AND rewrites the carousel config accordingly.
+	 * 
+	 * A set of carousel slides is represented as a map with values as tags; e.g.,
+	 * tag "0" for the first structure and so on, a highly unfortunate use of data
+	 * in the tags. I don't know how to build a POJO that maps to such a structure,
+	 * so parse here with basic Jackson features.
+	 */
+	private static String migrateCarouselConfig(CMSReaderClient cmsClient, ICommonDataServiceRestClient cdsClient,
+			String bgPath, String igPath, String keyPrefix, String jsonString)
+			throws JsonProcessingException, IOException {
+		ObjectMapper mapper = new ObjectMapper();
+		JsonNode configNode = mapper.readTree(jsonString);
+		for (int index = 0; true; ++index) {
+			JsonNode slide = configNode.get(Integer.toString(index));
+			if (slide == null)
+				break;
+			logger.info("Processing {} carousel slide {}", keyPrefix, index);
+			// Check if already migrated - presence of unique number
+			if (slide.get(uniqueKey) != null || slide.get(bgImgKey) != null || slide.get(infoImgKey) != null) {
+				logger.info("Already migrated, skipping");
+				continue;
+			}
+			if (!(slide instanceof ObjectNode)) {
+				logger.warn("Unexpected object type");
+				continue;
+			}
+			ObjectNode slideNode = (ObjectNode) slide;
+			slideNode.set(uniqueKey, new IntNode(index));
+
+			JsonNode bgNode = slideNode.get(bgImageUrl);
+			if (bgNode != null) {
+				String url = bgNode.asText();
+				logger.info("Migrating background image {}", url);
+				byte[] img = cmsClient.getCarouselImage(bgPath, url);
+				if (img == null || img.length == 0) {
+					logger.warn("Failed to get background image {}/{}", bgPath, url);
+				} else {
+					// build and add key for image
+					final String imgKey = keyPrefix + "." + Integer.toString(index) + ".bgImg";
+					slideNode.set(bgImgKey, new TextNode(imgKey));
+					MLPSiteContent content = new MLPSiteContent(imgKey, img, "image/jpg");
+					cdsClient.createSiteContent(content);
+					// remove old tag
+					slideNode.remove(bgImageUrl);
+				}
+			}
+			JsonNode igNode = slideNode.get(infoImageUrl);
+			if (igNode != null) {
+				String url = igNode.asText();
+				logger.info("Migrating infographic image {}", url);
+				byte[] img = cmsClient.getCarouselImage(igPath, url);
+				if (img == null || img.length == 0) {
+					logger.warn("Failed to get infographic image {}/{}", igPath, url);
+				} else {
+					// build and add key for image
+					final String imgKey = keyPrefix + "." + Integer.toString(index) + ".infoImg";
+					slideNode.set(bgImgKey, new TextNode(imgKey));
+					MLPSiteContent content = new MLPSiteContent(imgKey, img, "image/jpg");
+					cdsClient.createSiteContent(content);
+					// special case among special cases ARGH
+					slideNode.set(hasInfoGraphic, BooleanNode.getTrue());
+					// remove old tag
+					slideNode.remove(infoImageUrl);
+				}
+			}
+
+		} // for each slide
+		return mapper.writeValueAsString(configNode);
 	}
 
 	/**
 	 * Factors code of loop above
 	 * 
 	 * @param name
-	 *            Name of document
+	 *                 Name of document
 	 * @param docs
-	 *            List of documents
+	 *                 List of documents
 	 * @return True if doc with specified name occurs in list
 	 */
 	private static boolean findDocNameInCdsList(String name, List<MLPDocument> docs) {
@@ -308,11 +514,11 @@ public class MigrateCmsToCdsApp {
 	 * Forms Nexus group id as prefix.SolutionId.RevisionId
 	 * 
 	 * @param prefix
-	 *            Nexus prefix string
+	 *                       Nexus prefix string
 	 * @param solutionId
-	 *            Solution ID
+	 *                       Solution ID
 	 * @param revisionId
-	 *            Revision ID
+	 *                       Revision ID
 	 * @return Dotted string suitable for use as Nexus group id
 	 */
 	private static String createNexusGroupId(String prefix, String solutionId, String revisionId) {
@@ -325,7 +531,7 @@ public class MigrateCmsToCdsApp {
 	 * Splits name into basename and extension at the last period.
 	 * 
 	 * @param name
-	 *            containing a period
+	 *                 containing a period
 	 * @return Array of size 2; empty if the name has no period
 	 */
 	private static String[] splitFileBaseExt(String name) {