+++ docs/release-notes.rst
@@ -26,6 +26,7 @@ available as a Docker image in the Acumos/LF Docker registry.
 
 Version 2.1.1, 20 Feb 2019
 --------------------------
+* Add controller methods to get pages of configs and contents (`ACUMOS-2423 <https://jira.acumos.org/browse/ACUMOS-2423>`_)
 * Add peer-status code UK (`ACUMOS-2513 <https://jira.acumos.org/browse/ACUMOS-2513>`_)
 * Allow server to run as unprivileged user (`ACUMOS-2533 <https://jira.acumos.org/browse/ACUMOS-2533>`_)
 * Requires database schema version 2.1